Refactor project structure to make it more maintainable
